message php à afficher plus explicites, factoring de la fonction de display de messages dans un fichier php tierce, ajout d'anim de slide pour les toasts
