+++ src/index.php
@@ -3,7 +3,12 @@
 	require_once("./programmation/menu.php");
 	require_once("./database/dataBase.php");
 	
-	//~ error_reporting(E_ALL ^ E_WARNING);
+	/* 
+	 * uncomment if you don't want to show warning anymore
+	 * usefull for bdd connexion when bad credentials are given
+	 * 
+	 */
+	// error_reporting(E_ALL ^ E_WARNING);
 	
 	menu();
 ?>
